refactor: split `MockRequestIdServiceStub`
